requirements.txt
Flask==3.1.0
Flask-JWT-Extended == 4.7.1
pandas==2.2.3
openpyxl==3.1.5
mysql-connector-python==9.1.0
bcrypt == 3.2.0
python-dotenv==1.0.1
